[MRKT3] Making custom reticles a bit easier in MRTK3. (#11758)
## Overview

The intent of this change is to make attaching a custom reticle easier in MRTK3. The custom reticle is now attached to the base reticle's root transform, removing the need for the custom reticle to position itself.

Also, reverting changes made to `RecticleMagnetism`. With this change, `RecticleMagnetisms` once again uses an interface for setting progress, instead of requiring a `RingReticle` component.  Also, instead of using the old `IVariableReticle`, `RecticleMagnetisms` is now using `IVariableProgressReticle` to better indicate the propose of the interface.

I also rename the newer `IVariableReticle` to `IRecticleVisual` to better indicate its purpose.

This change is in response to https://github.com/microsoft/MixedRealityToolkit-Unity/discussions/11751

## Changes
- Fixes: https://github.com/microsoft/MixedRealityToolkit-Unity/discussions/11751