Update prefabs to no longer be variants
These were direct variants of the imported glTF models. Unfortunately, it appears that the imported glTF models don't keep a consistent fileID across glTF importer versions. This was causing the reference to the controller model prefabs to be lost in the hand prefabs when different glTF importer versions were used compared to the repo's sample scene.